refactor plugin tests
