+++ tests/contrib/tortoise_orm/test_tortoise_orm_plugin_openapi_specs.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 from starlite import Starlite
 from starlite.contrib.tortoise_orm import TortoiseORMPlugin
-from tests.plugins.tortoise_orm import (
+from tests.contrib.tortoise_orm import (
     create_tournament,
     get_tournament,
     get_tournaments,
