+++ tests/dto_factory/test_dto_factory_model_conversion.py
@@ -10,7 +10,7 @@ from starlite.contrib.tortoise_orm import TortoiseORMPlugin
 from starlite.dto import DTOFactory
 from starlite.exceptions import ImproperlyConfiguredException
 from tests import Person, TypedDictPerson, VanillaDataClassPerson
-from tests.plugins.tortoise_orm import Tournament
+from tests.contrib.tortoise_orm import Tournament
 
 
 def _get_attribute_value(model_instance: Any, key: str) -> Any:
