+++ tests/test_plugins.py
